change to jquery plugin; add to bundle
